+++ base/cxx/utility/utility/digit/digit.h
@@ -26,14 +26,14 @@ public:
      * @param bit 二进制字符串的位数(选填), 不足的则补0
      * @return 二进制字符串
      */
-    static std::string dec2bin(unsigned int dec, unsigned int bit = 0);
+    static std::string dec2bin(unsigned long long dec, unsigned int bit = 0);
 
     /**
      * @brief 二进制字符串转十进制数字
      * @param binStr 二进制字符串
      * @return 十进制数字
      */
-    static unsigned int bin2dec(const std::string& binStr);
+    static unsigned long long bin2dec(const std::string& binStr);
 
     /**
      * @brief 十进制数字转十六进制字符串
@@ -43,13 +43,13 @@ public:
      * @param isUpper 是否输出大写(选填), 默认大写
      * @return 十六进制字符串
      */
-    static std::string dec2hex(unsigned int dec, unsigned int bit = 0, bool prefix = false, bool isUpper = true);
+    static std::string dec2hex(unsigned long long dec, unsigned int bit = 0, bool prefix = false, bool isUpper = true);
 
     /**
      * @brief 十六进制字符串转十进制数字
      * @param hexStr 十六进制字符串
      * @return 十进制数字
      */
-    static unsigned int hex2dec(const std::string& hexStr);
+    static unsigned long long hex2dec(const std::string& hexStr);
 };
 } // namespace utility
